More libcurl-webrequest changes
- Added 'IWebRequestProxyOptions' interface (not implemented by 'CURLRequest' yet).
- Added  more functions to 'IWebRequestOptions' and 'IWebResponse' interfaces.
- Implemented handling of an unauthorized response.
- Use of 'HTTPStatus' type for response code is replaced by 'std::optional<int>' object because not all possible request use HTTP protocol.
- HTTPStatus can now be constructed from 'std::optional<T>' where T is an integral type.

+++ kxf/Network/CURLWebSession/CURLUtility.cpp
@@ -109,7 +109,7 @@ namespace kxf::CURL::Private
 	{
 		return DoDispatchSetOptionInt(m_Type, m_Handle, option, value);
 	}
-	bool SessionHandle::SetOption(int option, size_t value) noexcept
+	bool SessionHandle::SetOption(int option, uint64_t value) noexcept
 	{
 		return DoDispatchSetOptionInt(m_Type, m_Handle, option, value);
 	}
